improvements to boot sector
Although it was previously working on qemu, it was not working on real
hardware via a USB thumb drive. This commit adds zeroing of segment
registers, a disk time stamp and a partition table and now it works
on several PCs.

Almost all comments have been refined and a new print_string_16.asm
file is now here to provide a minimal import of a print function,
instead of the general and wasteful previous import of several
functions.